changed the server deployment everything in backend is working

+++ lib/widgets/cafeteria_card.dart
@@ -13,7 +13,7 @@ class Cafeteria {
   final String name;
   final String cafeId;
   final String description;
-  final double seats;
+  final int seats;
   List<LunchMenu> cafeMenuList = [];
   String imageURL =
       "https://images.unsplash.com/photo-1602016753527-3b369b88af5f?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=1170&q=80";
